+++ big_size/index.js
@@ -1,3 +1,4 @@
 const decodeBigSize = require('./decode_big_size');
+const encodeBigSize = require('./encode_big_size');
 
-module.exports = {decodeBigSize};
+module.exports = {decodeBigSize, encodeBigSize};