+++ index.js
@@ -1,3 +1,4 @@
 const {decodeBigSize} = require('./big_size');
+const {encodeBigSize} = require('./big_size');
 
-module.exports = {decodeBigSize};
+module.exports = {decodeBigSize, encodeBigSize};