+++ test/big_size/test_decode_big_size.js
@@ -9,44 +9,44 @@ const tests = [
     error: 'ExpectedHexEncodedBigSizeValueToDecode',
   },
   {
-    args: {encoded: '00'},
+    args: {encoded: '0001'},
     description: 'Zero value is decoded',
-    expected: {decoded: '0'},
+    expected: {decoded: '0', length: 1},
   },
   {
     args: {encoded: 'fc'},
     description: 'A single byte value is decoded',
-    expected: {decoded: '252'},
+    expected: {decoded: '252', length: 1},
   },
   {
     args: {encoded: 'fd00fd'},
     description: 'The smallest two byte number is decoded',
-    expected: {decoded: '253'},
+    expected: {decoded: '253', length: 3},
   },
   {
     args: {encoded: 'fdffff'},
     description: 'The largest two byte number is decoded',
-    expected: {decoded: '65535'},
+    expected: {decoded: '65535', length: 3},
   },
   {
     args: {encoded: 'fe00010000'},
     description: 'The smallest four byte number is decoded',
-    expected: {decoded: '65536'},
+    expected: {decoded: '65536', length: 5},
   },
   {
     args: {encoded: 'feffffffff'},
     description: 'The largest four byte number is decoded',
-    expected: {decoded: '4294967295'},
+    expected: {decoded: '4294967295', length: 5},
   },
   {
     args: {encoded: 'ff0000000100000000'},
     description: 'The smallest eight byte number is decoded',
-    expected: {decoded: '4294967296'},
+    expected: {decoded: '4294967296', length: 9},
   },
   {
     args: {encoded: 'ffffffffffffffffff'},
     description: 'The largest eight byte number is decoded',
-    expected: {decoded: '18446744073709551615'},
+    expected: {decoded: '18446744073709551615', length: 9},
   },
   {
     args: {encoded: 'fd00fc'},
@@ -100,9 +100,10 @@ tests.forEach(({args, description, error, expected}) => {
     if (!!error) {
       throws(() => decodeBigSize(args), new Error(error), 'Got error');
     } else {
-      const {decoded} = decodeBigSize(args);
+      const value = decodeBigSize(args);
 
-      equal(decoded, expected.decoded, 'Got expected result');
+      equal(value.decoded, expected.decoded, 'Got expected result');
+      equal(value.length, expected.length, 'Got expected length');
     }
 
     return end();