+++ CHANGELOG.md
@@ -1,5 +1,10 @@
 # Versions
 
+## 1.1.0
+
+- `decodeBigSize`: Add `length` attribute to indicate the BigSize byte length
+- `encodeBigSize`: Add method to encode BigSize type serialized numbers
+
 ## 1.0.0
 
 - `decodeBigSize`: Add method to decode BigSize type serialized numbers